separate github copilot from chatmodelmanager into new file
